requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-identity
# azure-identity-broker
azure-core
azure-functions
azure-ai-inference
langchain-azure-dynamic-sessions
openai
